requirements.txt
aiogram~=2.21
clean-text==0.6.0
asyncpg~=0.26.0
python-dotenv